+++ apps/webstat-main/src/controllers/base-controller.go
@@ -6,12 +6,12 @@ import (
 	"github.com/aws/aws-lambda-go/events"
 )
 
-func CreateResponse(statusCode int, body interface{}) (*events.APIGatewayProxyResponse, error) {
+func CreateResponse(statusCode int, body interface{}) (*events.APIGatewayV2HTTPResponse, error) {
 	headers := make(map[string]string)
 	headers["Content-Type"] = "application/json"
 
 	stringBody, _ := json.Marshal(body)
-	return &events.APIGatewayProxyResponse{
+	return &events.APIGatewayV2HTTPResponse{
 		Headers:    headers,
 		StatusCode: statusCode,
 		Body:       string(stringBody),
